# src/main/resources/openapi/api.yml
openapi: "3.0.2"
info:
  title: products_service
  description: Microservice that handle the lifecycle from products
  version: 1.0.0
paths:

  /products:
    get:
      tags:
        - products
      summary: Return the registered products
      operationId: getProducts
      responses:
        '200':
          description: A JSON array of user names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductsResponse'

    post:
      tags:
        - products
      summary: Create the product
      operationId: createProduct
      responses:
        '200':
          description: A JSON array of user names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProductResponse'
      requestBody:
        description: Information to create the product
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'


components:
  schemas:

    CreateProductRequest:
      description: Contain the basic information from the creation of the product
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          type: number
          format: double

    CreateProductResponse:
      description: Contain the id of the generated product
      type: object
      required:
        - data
      properties:
        data:
          type: string
    GetProductsResponse:
      description: Contain the registered products details
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductsDetail'

    ProductsDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: double